added hp and mana bar to battle print some fixes
