+++ TestingGrounds.py
@@ -3,7 +3,7 @@
 Use this space to test out features!
 """
 import Game
-from Class_NPC import NPC
+from old_Class_NPC import NPC
 from Class_Hero import Hero
 from Class_Party import Party
 from battle import *
